Create the initial `harvest_translate` bin crate.
This adds a mostly-empty crate for the `harvest_translate` binary, along with some basic test infrastructure for it (verify code builds, clippy, and formatting).

The name `harvest_translate` is definitely not final and is very bikeshedable. It's also possible that we'll just have a general `harvest` binary, of which `translate` is just one subcommand.