Introduce the `Tool` trait.
This is the trait the scheduler will use to evaluate whether it can run a particular tool and to invoke that tool.

There is a lot to be figured out: currently it is assumed that tools magically know the `Id`s of the representations they use, and it's not clear to me how that will work.