Tool trait: choose the "read snapshot, no conflicting writes" concurrency model.
Before running a tool, the scheduler will invoke `Tool::may_write` to learn which IR IDs the tool will write. The scheduler uses this to avoid running the tool if it will edit the same ID as another concurrent tool invocation. `may_write` has read-only access to a snapshot of the IR which it can scan through to find representations it is interested in.

If the tool is ran, it is given a `context` object that gives it read-only access to the same snapshot as `may_write` saw as well as write access to the IDs that `may_write` returned. The tool invocation can also create new IDs and write to those representations, as they are guarantee to not conflict with other tool invocations.

`context`'s interface is not yet implemented, because implementing it will require making a number of changes to the IR that will conflict with another PR.