Make `Id`'s tests parallel safe.
This refactors the implementation of `Id::new_array` so that its unit test can inject a different `AtomicU64`. This allows the test to pass even if a different test concurrently calls `Id::new_array`. A new test is added to verify that `Id::new` and `Id::new_array` correctly use the refactored implementation.