Implement configuration using the `config` crate.
This loads configuration from three places, in increasing order of precedence:

1. A default configuration file.
2. A per-user config file (the location depends on the OS and can be printed using --print-config-path).
3. A new `--config` command line argument.

I've realized that the configuration and command line interface have a lot of overlap and probably belong together. Calling the module `config` seems sensible because I expect the configuration to be larger than the CLI, but there are too many `config`s in the code, so I kept the module name (`cli`) as-is.

I migrated from `tempdir` to `tempfile`; `tempdir` is deprecated in favor of `tempfile` and `tempfile` allows us to set the permissions on the temporary directory.