Implement a more complete scheduler + tool invoker.
The existing scheduler was an MVP to make the raw_source_to_cargo_llm tool execute. This splits a lot of the non-scheduling-related functionality out of Scheduler into new objects. It makes several improvements to functionality:
1. Tools are now invoked in their own threads, and currently-in-use IDs are tracked correctly.
2. Tool panics are now caught and handled as errors rather than terminating full program execution.
3. The main loop now saves proposed tool invocations if they are not currently runnable and tries to run them again later (the next step towards a proper useful scheduler).
4. The main loop is moved into `main()`. Having the main loop in `Scheduler` tends to make all the program's state live in Scheduler which is not great.

Items 1 + 2 are useful to do now, as they will verify that everything that needs to be `Send` and `UnwindSafe` are.