Introduce `harvest_ir::edit::Organizer`.
Currently, there are a couple of things about managing IR edits that are messy:

1. `Edit::new` is public, so tools can create `Edit`s and swap them around.
2. Tools can return IDs that are not part of the IR from `Tool::might_write`, which is a logic error.

In both cases, `harvest_translate` *should* be able to catch the error and emit a diagnostic, but when I went to implement it there's just no sensible place for that logic.

Instead, this introduces a new component whose main purpose is to create and apply `Edit`s, `edit::Organizer`. Because it is in the `harvest_ir` crate, `Edit::new` can be removed, so tools can no longer create `Edit`s at will. It also provides a good place to verify that `Tool::might_write` only returned IDs that are in the IR.

I also moved the IR storage out of the scheduler, it doesn't really belong there.