Better handling of a non-empty output directory.
This makes harvest_translate error on startup if the output directory is non-empty (by default). Optionally, it can be configured to empty the contents of the output directory (this is so you can test harvest_translate by hand without having to clear out the output directory every time).

I put the logic for this into a new `util` module because I plan to re-use it for the diagnostics directory.