HarvestIR::get_by_representation fixes + related changes.
This fixes HarvestIR::get_by_representation so that it works correctly, and adds a unit test verifying that. It also changes get_by_representation to return both the IDs and the representations.

I also added HarvestIR::add_representation (following the same API as Edit::add_representation) to make the unit test easier to write.