Diagnostics: wait for reporters to be dropped before returning Diagnostics.
This changes the ownership model of `diagnostics::Shared`. Now, it is owned by an `Arc`. When `Collector::diagnostics` is called, the `Collector`'s handle is dropped, which (if the other handles have already been dropped), will result in the `Shared` being dropped. When `Shared` is dropped, it sends its `Diagnostics` through a channel to the `Collector`, which the `Collector` waits on.

If we later decide that we don't want the diagnostics system to wait for reporters to be dropped, we can change the other `Shared` references into `Weak` references.