Diagnostics design: use tool names for per-tool directories.
The current design in the design doc -- making the tool directory name equal to the number of the resulting IR revision -- has a few problems:

1. If a tool fails, it does not produce an IR revision, but should still have a diagnostic directory.
2. The directory name cannot be determined until the tool finishes, which means we need to rename the directory after tool completion. This will cause issues anywhere we store paths to files in a tool's directory.

This changes the tool directory names to consist of the tool name and per-tool run number. This allows the directory name to be determined when the tool is launched and remain constant.

Because snake-case is the norm for file and directory names, this changes the tool names to snake case.

I also tweaked the design doc examples to use 3-wide count numbers, because that is the field size I plan to use at first (based on a guess at how many tool invocations `translate` will eventually do).