Tool diagnostics reporting framework.
This adds ToolReporter and other associated types, which tools will use to record diagnostics. This also changes the behavior of the tool runner to wait for all tool threads to terminate before considering a tool's execution complete.

This PR doesn't implement many new diagnostics, that will be done in future PRs.