Refactor tools to separate crates
This is a big modification. On the surface it just moves the tools and
representations into separate crates, but this requires non-trivial
structural changes, especially to how configuration is handled to avoid
circular dependencies.

One notable design choice is to put CargoPackage and RawSrc
representations into the harvest_ir crate, since they are likely broadly
useful building blocks rather than a relevent mostly to a specific tool.