Make `reporter` a generic closure
This aims to decouple RunContext (and thus Tool) from the specific
internals of the translate/diagnostics libraries. Rather than store a
ToolReporter directly in a the RunContext struct, RunContext stores an
arbitrary closure that yields a value to drop at when a thread
completes.

Also introduces a wrapper around `std::thread::spawn` that generates and
drops the guard.